Post-review changes
- Alphabeticises input kwargs to Combine plugin
- Moves filter_realizations to cube_manipulation.py
- Improves doc-strings
- Improves error message
- Reorders if/else to be positive instead of negative
